abxLS refactoring part2
