+++ zerospeech/benchmarks/_benchmarks/abx_LS/leaderboard.py
@@ -1,22 +1,21 @@
-from typing import List
-
-from pydantic import BaseModel
+from typing import List, Literal
 
 from zerospeech.model import m_leaderboard
 
 
-class ABXLSScoresSet(BaseModel):
-    dev: m_leaderboard.ABXScoreTuple
-    test: m_leaderboard.ABXScoreTuple
-
-
-class ABXLSScores(m_leaderboard.LeaderboardScores):
-    clean: ABXLSScoresSet
-    other: ABXLSScoresSet
+class ABXLSScore(m_leaderboard.LeaderboardScores):
+    subset: Literal['dev-clean', 'dev-other', 'test-clean', 'test-other']
+    granularity: Literal['triphone', 'phoneme']
+    speaker_mode: Literal['across', 'within']
+    context_mode: Literal['within', 'any']
+    score: float
+    item_file: str
+    pooling: str
+    seed: int
 
 
 class ABXLSEntry(m_leaderboard.LeaderboardEntry):
-    scores: ABXLSScores
+    scores: List[ABXLSScore]
 
 
 class ABXLS(m_leaderboard.Leaderboard):
