+++ zerospeech/benchmarks/tasks/abx/abx_phoneme/__init__.py
@@ -1,0 +1,2 @@
+from .task import SimpleABXPhonemeTask
+from .params import ABXSpeakerMode, ABXDistanceMode, ContextMode, ABX2Parameters
