--- zerospeech/benchmarks/tasks/abx_phoneme/__init__.py
@@ -1,2 +1,0 @@
-from .task import SimpleABXPhonemeTask
-from .params import ABXMode, ABXDistanceMode, ContextMode, ABX2Parameters
